vertex_agent.py
# vertex_agent.py
import os
from dotenv import load_dotenv
import vertexai
from vertexai.generative_models import GenerativeModel

load_dotenv()  # Load environment variables from .env

PROJECT_ID = os.getenv("GCP_PROJECT_ID")
LOCATION = os.getenv("GCP_LOCATION")

vertexai.init(project=PROJECT_ID, location=LOCATION)
model = GenerativeModel("gemini-1.5-pro")

def ask_llm(prompt: str) -> str:
    response = model.generate_content(prompt)
    return response.text
